requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
# Choose one of the following MySQL libraries
# Flask-MySQLdb==2.0.0
mysql-connector-python==9.0.0
# mysqlclient==2.2.4
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
gunicorn==20.1.0  # Add this 